move checking to getAllStatefulsets
